Update ch-mount-options.rst note
I updated the note to take a better consideration of what mount options are file-system specific and which work for each sub-volume. There is a better explanation in a different section of the manual (ch-subvolume-intro.rst). 
This removes the note and instead copies the text over from that section. 

I think this explains the information a bit better.