edited note for mounting options
ch-subvolume-intro.rst has a much better explanation of which mount options are generic vs which are file system based. I have replaced the note with the explanation from ch-subvolume-intro.rst

This refers the user to look somewhere else for generic options while giving a very small explanation of what the generic options are. 

This can prevent confus